separating menu items, checkbox styling
