+++ src/main/frontend/src/components/shopPage/CategoriesMenuItem.module.css
@@ -1,0 +1,38 @@
+.collapsibleCategory {
+    width: 260px;
+    background-color: #FFF;
+    border: none;
+    text-align: left;
+    padding: 12px 24px;
+    font-size: 16px;
+}
+
+.collapsibleCategory:hover {
+    color: #8367D8;
+}
+
+.collapseIcon {
+    float: right;
+}
+
+.subcategoriesContainer {
+    text-align: left;
+}
+
+.subcategoryItem {
+    padding: 15px 28px;
+}
+
+.subcategoryCheckbox {
+    appearance: none;
+    cursor: pointer;
+    border: 1px solid #8367D8;
+    height: 15px;
+    width: 15px;
+    margin-right: 12.5px;
+    border-radius: 2px;
+}
+
+.subcategoryCheckbox:checked {
+    background-color: #8367D8;
+}
